# src/main/resources/static/openapi/exercise-service.yaml
openapi: 3.1.0
info:
  title: Exercise Service API
  version: 1.0.0
  description: Single-user exercise log book.
servers:
  - url: /
paths:
  /api/v1/exercise-logs:
    post:
      operationId: createExerciseLog
      summary: Create an exercise log
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateExerciseLogRequest' }
      responses:
        '201':
          description: Exercise log created
          headers:
            Location: { required: true, schema: { type: string, format: uri-reference } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExerciseLog' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
    get:
      operationId: listExerciseLogs
      summary: List exercise logs
      parameters:
        - name: exerciseName
          in: query
          schema: { type: string, minLength: 1 }
          description: Case-insensitive exercise-name filter.
        - name: sort
          in: query
          schema:
            type: string
            enum: [loggedAt, highestSetVolume]
            default: loggedAt
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        '200':
          description: Exercise logs
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/ExerciseLog' } }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/v1/exercise-logs/exercises:
    get:
      operationId: listDistinctExercises
      summary: List distinct exercises
      description: Returns unique exercise names with log counts and latest log timestamps.
      responses:
        '200':
          description: Distinct exercises
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/ExerciseSummary' } }
  /api/v1/exercise-logs/{exerciseLogId}:
    get:
      operationId: getExerciseLog
      summary: Get one exercise log
      parameters:
        - name: exerciseLogId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Exercise log
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExerciseLog' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
components:
  schemas:
    CreateExerciseLogRequest:
      type: object
      additionalProperties: false
      required: [exerciseName, sets]
      properties:
        exerciseName:
          type: string
          minLength: 1
          maxLength: 120
          pattern: '.*\S.*'
        loggedAt: { type: string, format: date-time }
        sets:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/ExerciseSetRequest' }
    ExerciseSetRequest:
      type: object
      additionalProperties: false
      required: [setNumber, weightKg, reps]
      properties:
        setNumber: { type: integer, minimum: 1, maximum: 32767 }
        weightKg: { type: number, minimum: 0, multipleOf: 0.01 }
        reps: { type: integer, minimum: 1, maximum: 32767 }
    ExerciseLog:
      type: object
      additionalProperties: false
      required: [id, exerciseName, loggedAt, sets, highestSetVolumeKg]
      properties:
        id: { type: integer, format: int64 }
        exerciseName: { type: string }
        loggedAt: { type: string, format: date-time }
        sets: { type: array, items: { $ref: '#/components/schemas/ExerciseSet' } }
        highestSetVolumeKg:
          type: number
          description: Highest single-set volume, calculated as weightKg multiplied by reps.
    ExerciseSet:
      type: object
      additionalProperties: false
      required: [id, setNumber, weightKg, reps, volumeKg]
      properties:
        id: { type: integer, format: int64 }
        setNumber: { type: integer }
        weightKg: { type: number }
        reps: { type: integer }
        volumeKg:
          type: number
          description: Calculated as weightKg multiplied by reps.
    ExerciseSummary:
      type: object
      additionalProperties: false
      required: [exerciseName, logCount, lastLoggedAt]
      properties:
        exerciseName: { type: string }
        logCount: { type: integer, minimum: 1 }
        lastLoggedAt: { type: string, format: date-time }
    Problem:
      type: object
      required: [type, title, status, detail, code]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        code: { type: string }
  responses:
    BadRequest:
      description: Invalid request
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    Conflict:
      description: Request conflicts with persisted data
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    NotFound:
      description: Exercise log not found
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }